services/reploop-recommender-llm/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.7.0
ollama>=0.3.0
